Move iroh discovery to iroh_host
